+++ src/features/personal-my-page-feature/_components/edit-profile/index.tsx
@@ -1,5 +1,5 @@
 import { EditProfileCss } from './indexCss';
-import { SubmitButton } from '../button';
+import { SubmitButton } from '@/components/button';
 import InputWithLabel from './input-with-label/InputWithLabel';
 import EditProfileImg from './edit-profile-img/EditProfileImg';
